Silence more 'comparison between signed and unsigned' warnings.

+++ Include/pymem.h
@@ -90,10 +90,10 @@ PyAPI_FUNC(void) PyMem_Free(void *);
  */
 
 #define PyMem_New(type, n) \
-  ( ((n) > PY_SSIZE_T_MAX / sizeof(type)) ? NULL : \
+  ( ((size_t)(n) > PY_SSIZE_T_MAX / sizeof(type)) ? NULL :	\
 	( (type *) PyMem_Malloc((n) * sizeof(type)) ) )
 #define PyMem_NEW(type, n) \
-  ( ((n) > PY_SSIZE_T_MAX / sizeof(type)) ? NULL : \
+  ( ((size_t)(n) > PY_SSIZE_T_MAX / sizeof(type)) ? NULL :	\
 	( (type *) PyMem_MALLOC((n) * sizeof(type)) ) )
 
 /*
@@ -103,10 +103,10 @@ PyAPI_FUNC(void) PyMem_Free(void *);
  * caller's memory error handler to not lose track of it.
  */
 #define PyMem_Resize(p, type, n) \
-  ( (p) = ((n) > PY_SSIZE_T_MAX / sizeof(type)) ? NULL : \
+  ( (p) = ((size_t)(n) > PY_SSIZE_T_MAX / sizeof(type)) ? NULL :	\
 	(type *) PyMem_Realloc((p), (n) * sizeof(type)) )
 #define PyMem_RESIZE(p, type, n) \
-  ( (p) = ((n) > PY_SSIZE_T_MAX / sizeof(type)) ? NULL : \
+  ( (p) = ((size_t)(n) > PY_SSIZE_T_MAX / sizeof(type)) ? NULL :	\
 	(type *) PyMem_REALLOC((p), (n) * sizeof(type)) )
 
 /* PyMem{Del,DEL} are left over from ancient days, and shouldn't be used
